Move tracking step into its own file.
